The robot schedule_monitor receives the d_graph_update_cb, not the executor_interface
